requirements.txt
boto3
streamlit
langchain
faiss-cpu
pypdf
logging
python-dotenv
